캐릭터에서 하던 이동가능범위처리를 씬으로 옮김 깔끔하게 맵밖으로 나가지 못하게 함 주석처리

+++ D2DExample/D2D/MainMenuScene.cpp
@@ -3,8 +3,8 @@
 
 CMainMenuScene::CMainMenuScene(void)
 {
-	float width = NNApplication::GetInstance()->GetScreenWidth();
-	float height = NNApplication::GetInstance()->GetScreenHeight();
+	float width = (float)NNApplication::GetInstance()->GetScreenWidth();
+	float height = (float)NNApplication::GetInstance()->GetScreenHeight();
 	m_LogoLabelEnglish = NNLabel::Create( L"JuGums", L"Something Strange", 100.f );
 	m_LogoLabelEnglish->SetColor( 255, 0, 0 );
 	m_LogoLabelEnglish->SetPosition( width/2 - 25.f, height/4 - 100.f );
